Move web models to new module
